Cleanup temp CF_HOME

+++ src/staticfile/integration/integration_suite_test.go
@@ -5,6 +5,7 @@ import (
 	"flag"
 	"fmt"
 	"os"
+	"strings"
 	"time"
 
 	"github.com/cloudfoundry/libbuildpack/cutlass"
@@ -59,7 +60,9 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 
 var _ = SynchronizedAfterSuite(func() {
 	// Run on all nodes
-	Expect(os.RemoveAll(os.Getenv("CF_HOME"))).To(Succeed())
+	if strings.HasPrefix(os.Getenv("CF_HOME"), "/tmp") {
+		Expect(os.RemoveAll(os.Getenv("CF_HOME"))).To(Succeed())
+	}
 }, func() {
 	// Run once
 	Expect(cutlass.RemovePackagedBuildpack(packagedBuildpack)).To(Succeed())
